+++ src/eninput.c
@@ -7,7 +7,7 @@
 #include "shared/common.h"
 
 void waitForKeyOrJoy(void) {
-    int key = 0; /* joystick-button exit leaves key unset; 0 = don't treat as Alt-Q */
+    int16 key = 0; /* joystick-button exit leaves key unset; 0 = don't treat as Alt-Q */
     if (commData->setupUseJoy == 1) {
         do {
             if (misc_checkKeyBuf() == 0) {
