+++ src/shared/filepic.c
@@ -7,16 +7,16 @@
 #include "../log.h"
 #include <SDL3/SDL.h>
 
-SDL_IOStream *openFile(const char *name, int mode);
+SDL_IOStream *openFile(const char *name, int16 mode);
 void fileClose(SDL_IOStream *handle);
-void decodePic(SDL_IOStream *handle, int segment);
+void decodePic(SDL_IOStream *handle, int16 segment);
 
 void mystrcpy(char *dest, const char *source) {
     do {
     } while ((*dest++ = *source++) != '\0');
 }
 
-SDL_IOStream *openFileWrapper(const char *filename, int mode) /* Original: OpenFile(file, attrib). */
+SDL_IOStream *openFileWrapper(const char *filename, int16 mode) /* Original: OpenFile(file, attrib). */
 {
     return openFile(filename, mode);
 }
@@ -26,7 +26,7 @@ void closeFileWrapper(SDL_IOStream *handle) /* Original: CloseFile(fh). */
     fileClose(handle);
 }
 
-void openShowPic(const char *filename, int page) /* Original chain: OpenFile + show/decode + CloseFile. Open, draw PIC to page, then close. */
+void openShowPic(const char *filename, int16 page) /* Original chain: OpenFile + show/decode + CloseFile. Open, draw PIC to page, then close. */
 {
     SDL_IOStream *fileHandle;
     fileHandle = openFileWrapper(filename, 0);
@@ -34,7 +34,7 @@ void openShowPic(const char *filename, int page) /* Original chain: OpenFile + s
     closeFileWrapper(fileHandle);
 }
 
-void loadPic(const char *filename, int segment) { /* Original chain: OpenFile + DecodePic(InSeg, OutSeg) + CloseFile. Load PIC into segment. */
+void loadPic(const char *filename, uint16 segment) { /* Original chain: OpenFile + DecodePic(InSeg, OutSeg) + CloseFile. Load PIC into segment. */
     SDL_IOStream *handle;
     handle = openFileWrapper(filename, 0);
     decodePic(handle, segment);
