+++ src/stalloc.h
@@ -3,6 +3,6 @@
 /* buffer alloc (stalloc.c) */
 #include "inttype.h"
 
-void *allocBuffer(int size);
+void *allocBuffer(int16 size);
 
 #endif /* F15_SE2_STALLOC */
